+++ expect/_matchers.ts
@@ -16,7 +16,7 @@ import { equal } from "./_equal.ts";
 import { format } from "./_format.ts";
 import { getMockCalls } from "./_mock_util.ts";
 import { inspectArg, inspectArgs } from "./_inspect_args.ts";
-import { iterableEquality } from "./_jest_utils.ts";
+import { iterableEquality } from "./iterable_equality.ts";
 import { AnyConstructor, MatcherContext, MatchResult } from "./_types.ts";
 import { buildEqualOptions } from "./_utils.ts";
