+++ expect/mod.ts
@@ -85,3 +85,4 @@
  */
 export { expect } from "./expect.ts";
 export { fn } from "./fn.ts";
+export { iterableEquality } from "./iterable_equality.ts";
